requirements/common.txt
celery==5.4.0
colorama==0.4.6
django==5.1.2
django-axes==7.0.0
django-admin-honeypot-updated-2021==1.2.0
django-admin-csvexport==2.2
django-import-export==4.1.1
django-taggit==6.1.0
django-extensions==3.2.3
django-celery-beat==2.7.0
django-cors-headers==4.3.1
django-celery-email==3.0.0 
django-debug-toolbar==4.4.2
django-filter==24.3
django-hosts==6.0
django-multiselectfield==0.1.13
django-jazzmin==3.0.1
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
django-redis==5.4.0
django-session-timeout==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
nanoid==2.0.0
phonenumbers==8.13.44
pillow==11.1.0
pip-chill==1.0.3
flower==2.0.1
python-dotenv==1.0.1
pygraphviz==1.14
redis==5.1.1
requests==2.32.3
whitenoise==6.7.0
boto3==1.35.76
django-storages==1.14.4
argon2-cffi==23.1.0
watchfiles==1.0.3
django-countries==7.6.1
django-autoslug==1.9.9
bs4==0.0.2
django-tinymce==4.1.0
Markdown==3.8
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
